Run konnect tests weekly and for dependabot PRs (#7630)
* tests add npm token to dependabot for prive pacakges

* tests(konnect) add konnect tests to release
add npm token for dependabot private package updates

* tests(konnect) add konnect retry

* tests(konnect) change koko test execution logic

* tests(konnect) fix missing env var logic

* tests(konnect) run workflow weekly and when there is package change

* tests(konnect) modify beforeAll hook for container removal

---------

+++ spec-ee/kong-api-tests/support/exec/gateway-container.ts
@@ -225,10 +225,16 @@ export const deployKonnectDataPlane = (controlPlaneEndpoint, telemetryEndpoint, 
  * @param {string} containerName 
  */
 export const stopAndRemoveTargetContainer = (containerName) => {
+  // if konnect-dp1 container exists
+  const doesContainerExists = execSync(`docker ps -a -q -f name=${containerName}`).toString().trim()
+  if (doesContainerExists) {
     try {
       execSync(`docker stop ${containerName}; docker rm ${containerName} -f`, { stdio: 'inherit' });
       console.info(`Successfully removed the ${containerName} docker container`)
     } catch (error) {
       console.error(`Something went wrong while removing the ${containerName} docker container`, error);
     }
+  } else {
+    console.info(`Konnect container ${containerName} doesn't exist, moving on with the rest of the test setup`)
   }
+}
